Moving lifecycle callback to separate file. Using isDebug() to switch Adjust environments.
