+++ app/src/main/java/mozilla/lockbox/support/AdjustLifecycleCallbacks.kt
@@ -1,0 +1,32 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package mozilla.lockbox.support
+
+import android.app.Activity
+import android.app.Application
+import android.os.Bundle
+import com.adjust.sdk.Adjust
+
+class AdjustLifecycleCallbacks : Application.ActivityLifecycleCallbacks {
+    override fun onActivityResumed(activity: Activity) {
+        Adjust.onResume()
+    }
+
+    override fun onActivityPaused(activity: Activity) {
+        Adjust.onPause()
+    }
+
+    override fun onActivityCreated(activity: Activity?, savedInstanceState: Bundle?) {}
+
+    override fun onActivityStarted(activity: Activity?) {}
+
+    override fun onActivityDestroyed(activity: Activity?) {}
+
+    override fun onActivitySaveInstanceState(activity: Activity?, outState: Bundle?) {}
+
+    override fun onActivityStopped(activity: Activity?) {}
+}
